+++ src/main/java/br/com/live/entity/DemandaProdutoPlano.java
@@ -4,7 +4,7 @@ public class DemandaProdutoPlano extends DemandaProduto {
 
 	public int plano;
 
-	public DemandaProdutoPlano(int id, String nivel, String grupo, String sub, String item, Double quantidade,
+	public DemandaProdutoPlano(int id, String nivel, String grupo, String sub, String item, int quantidade,
 			int periodo, int plano, int colecao, int linha, int artigo, int artigoCotas, int origem, int permanente,
 			int natureza, int nrInterno, int pedido, String embarque, int situacaoVenda) {
 		super(id, nivel, grupo, sub, item, quantidade, periodo, colecao, linha, artigo, artigoCotas, origem, permanente,
